requirements.txt
Flask
flask-oauthlib
azure-identity
python-dotenv
opencv-python